requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.0.0
numpy>=1.21.0
tqdm>=4.62.0